fix: remarks from review
